some correction in ex6.c

+++ loop_ex14.c
@@ -1,3 +1,9 @@
+/*
+*	Iteration Control Structure => Series exercise
+*	Behzod Hamrayev
+*/
+
+
 #include <stdio.h>
 
 int main()
@@ -8,11 +14,13 @@ int main()
 
 	float sum = 0; // natural sonlar yig'indisi
 	int k = 1; // oshuvchi o'zgaruvchi
+	// computation
 	while( sum < n )
 	{
-		sum += ((float)1/(float)k);
+		sum += (1.0/k);
 		k++;
 	}
+	// output result
 	printf( "k = %d, sum = %f\n", k, sum);
 
 	return 0;
